refactor #13: better structure for commmands logic
